requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==3.1.1
Werkzeug==3.1.3

# Flask Packages
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Security-Too==5.5.2
Flask-Babel==4.0.0
Flask-Mail==0.10.0
Flask-WTF==1.2.2
flask-session==0.8.0
email_validator==2.2.0
python-dotenv          # reads .flaskenv so FLASK_APP is set automatically
setuptools<81  # passlib requires pkg_resources, removed in setuptools 81+
argon2-cffi    # password hashing backend for Flask-Security-Too / passlib

# Automated tests
pytest==8.3.5
pytest-cov==6.1.0

# Run
gunicorn==23.0.0
